requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database dependencies
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.0

# Pydantic for data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LlamaIndex dependencies
llama-index-core>=0.10.0
llama-index-embeddings-openai>=0.1.0
llama-index-vector-stores-postgres>=0.1.0

# Scientific computing
numpy>=1.24.0

# Optional: for pgvector support
pgvector>=0.2.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0

# Environment management
python-dotenv>=1.0.0